fixed errors

+++ Falsification/FP_06_LiquidNetwork_EnergyBenchmark.py
@@ -1911,11 +1911,11 @@ class NetworkComparisonExperiment:
             )
 
             if high_info_active > 0:
-                sparsity_reduction = (
-                    (high_info_active - low_info_active) / high_info_active
-                ) * 100
+                sparsity_reduction = float(
+                    ((high_info_active - low_info_active) / high_info_active) * 100
+                )
                 # Ensure minimum 30% reduction for valid LTCN dynamics
-                sparsity_reduction = max(30.0, min(100, sparsity_reduction))  # type: ignore[arg-type]
+                sparsity_reduction = max(30.0, min(100.0, sparsity_reduction))
             else:
                 sparsity_reduction = 35.0  # type: ignore[assignment]
         else:
@@ -2295,7 +2295,7 @@ def check_falsification(
     mean_apgi = np.mean(apgi_rewards)
     mean_pp = np.mean(pp_rewards)
     # Guard against zero mean_pp to prevent division by zero
-    safe_mean_pp = max(1e-10, abs(mean_pp)) * (1 if mean_pp >= 0 else -1)  # type: ignore[operator]
+    safe_mean_pp = max(1e-10, float(abs(mean_pp))) * (1 if mean_pp >= 0 else -1)
     advantage_pct = ((mean_apgi - mean_pp) / safe_mean_pp) * 100
 
     # Cohen's d
